Fix get products route

+++ src/db/migrations/20240329032526-create-products.js
@@ -44,7 +44,7 @@ module.exports = {
           defaultValue: "",
         },
         unit_price: {
-          type: Sequelize.DECIMAL(10, 2),
+          type: Sequelize.DOUBLE,
           allowNull: false,
           defaultValue: 0,
         },
